refactor: move jobs inside main.py to its own domain
